requirements.txt
rembg==2.0.65
pillow==11.1.0
streamlit==1.44.1
streamlit-cropper==0.2.2